third_party/blink/web_tests/platform/mac-mac15-arm64/fast/webgl/context-gc-custom-properties-expected.txt

CONSOLE ERROR: Uncaught Unable to fetch WebGL rendering context for Canvas
Verify that the custom properties on a WebGL rendering context object are retained across GCs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught Unable to fetch WebGL rendering context for Canvas
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
